add parts page
This page is currently hard to reach since there's no path to it from
the top level.  You can get to it from a link on the horn history
page, though.  We need to decide whether to split the "Parts/Service"
tab into two tabs now that we've got the nice dynamic service locator.